upgrade lotus to v3
