+++ .changeset/thick-bats-tap.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/lotus-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
