--- packages/sources/lotus-test/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@chainlink/lotus-test-adapter",
-  "version": "0.0.0",
-  "description": "Chainlink lotus-test adapter.",
-  "keywords": [
-    "Chainlink",
-    "LINK",
-    "blockchain",
-    "oracle",
-    "lotus-test"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "url": "https://github.com/smartcontractkit/external-adapters-js",
-    "type": "git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "tsc -b",
-    "server": "node -e 'require(\"./index.js\").server()'",
-    "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
-    "start": "yarn server:dist"
-  },
-  "devDependencies": {
-    "@types/jest": "27.5.2",
-    "@types/node": "16.11.51",
-    "nock": "13.2.9",
-    "typescript": "5.0.4"
-  },
-  "dependencies": {
-    "@chainlink/external-adapter-framework": "0.30.2",
-    "ethers": "^5.4.7",
-    "tslib": "2.4.1"
-  }
-}
