--- packages/sources/lotus-test/src/endpoint/balance.ts
@@ -1,39 +1,0 @@
-import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
-import { InputParameters } from '@chainlink/external-adapter-framework/validation'
-import { config } from '../config'
-import { customTransport } from '../transport/balance'
-
-export const inputParameters = new InputParameters({
-  addresses: {
-    aliases: ['result'],
-    array: true,
-    type: {
-      address: {
-        type: 'string',
-        description: 'an address to get the balance of',
-        required: true,
-      },
-    },
-    description:
-      'An array of addresses to get the balances of (as an object with string `address` as an attribute)',
-    required: true,
-  },
-})
-
-export type BaseEndpointTypes = {
-  Parameters: typeof inputParameters.definition
-  Response: {
-    Data: {
-      result: string
-    }
-    Result: string
-  }
-  Settings: typeof config.settings
-}
-
-export const endpoint = new AdapterEndpoint({
-  name: 'balance',
-  aliases: ['Filecoin.WalletBalance'],
-  transport: customTransport,
-  inputParameters,
-})
