--- packages/sources/lotus-test/test-payload.json
@@ -1,6 +1,0 @@
-{
- "requests": [{
-  "from": "BTC",
-  "to": "USD"
- }]
-}
