+++ packages/sources/lotus/package.json
@@ -29,19 +29,15 @@
     "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
     "start": "yarn server:dist"
   },
-  "dependencies": {
-    "@chainlink/ea-bootstrap": "workspace:*",
-    "@chainlink/json-rpc-adapter": "workspace:*",
-    "ethers": "^5.4.7",
-    "tslib": "^2.3.1"
-  },
   "devDependencies": {
-    "@chainlink/ea-test-helpers": "workspace:*",
     "@types/jest": "27.5.2",
     "@types/node": "16.11.51",
-    "@types/supertest": "2.0.12",
     "nock": "13.2.9",
-    "supertest": "6.2.4",
     "typescript": "5.0.4"
+  },
+  "dependencies": {
+    "@chainlink/external-adapter-framework": "0.31.0",
+    "ethers": "^5.4.7",
+    "tslib": "2.4.1"
   }
 }
