--- packages/sources/lotus/schemas/env.json
@@ -1,21 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/lotus-adapter.json",
-  "title": "Chainlink Lotus Composite Adapter",
-  "description": "An external adapter to interact with the Lotus node API",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "Your Lotus node [API key/token](https://docs.filecoin.io/build/lotus/api-tokens/#obtaining-tokens)"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    },
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/json-rpc-adapter.json"
-    }
-  ]
-}
