+++ packages/sources/lotus/src/index.ts
@@ -1,10 +1,13 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute, endpointSelector } from './adapter'
-import * as endpoints from './endpoint'
-import { makeConfig, NAME } from './config'
-import * as types from './endpoint'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { config } from './config'
+import { balance } from './endpoint'
+import { PoRAdapter } from '@chainlink/external-adapter-framework/adapter/por'
 
-const adapterContext = { name: NAME }
+export const adapter = new PoRAdapter({
+  defaultEndpoint: balance.name,
+  name: 'LOTUS',
+  config,
+  endpoints: [balance],
+})
 
-const { server } = expose(adapterContext, makeExecute(), undefined, endpointSelector)
-export { NAME, makeExecute, makeConfig, server, types, endpoints }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
