--- packages/sources/lotus/test/integration/README.md
@@ -1,1 +1,0 @@
-See `packages/source/curve/test/integration` and `packages/composite/synth-index/test/integration` for reference
