+++ packages/sources/lotus/test/integration/adapter.test.ts
@@ -12,8 +12,10 @@ describe('execute', () => {
 
   beforeAll(async () => {
     oldEnv = JSON.parse(JSON.stringify(process.env))
-    ;(process.env.RPC_URL = process.env.RPC_URL ?? 'http://127.0.0.1:1234/rpc/v0'),
+    ;(process.env.FILECOIN_RPC_URL =
+      process.env.FILECOIN_RPC_URL ?? 'http://127.0.0.1:1234/rpc/v0'),
       (process.env.API_KEY = process.env.API_KEY ?? 'fake-api-key')
+    process.env['BACKGROUND_EXECUTE_MS'] = '0'
     const mockDate = new Date('2001-01-01T11:11:11.111Z')
     spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
