--- packages/sources/lotus/test/unit/balance.test.ts
@@ -1,41 +1,0 @@
-import { AdapterError, Requester } from '@chainlink/ea-bootstrap'
-import { assertError, setEnvVariables } from '@chainlink/ea-test-helpers'
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import { makeExecute } from '../../src'
-import process from 'process'
-import { TInputParameters } from '../../src/endpoint'
-
-let oldEnv: NodeJS.ProcessEnv
-
-beforeAll(() => {
-  oldEnv = JSON.parse(JSON.stringify(process.env))
-  process.env.CACHE_ENABLED = 'false'
-  process.env.API_KEY = process.env.API_KEY || 'test_key'
-})
-
-afterAll(() => {
-  setEnvVariables(oldEnv)
-})
-
-describe('execute', () => {
-  const jobID = '1'
-  const execute = makeExecute()
-
-  describe('validation error', () => {
-    const requests = [
-      { name: 'empty body', testData: {} },
-      { name: 'missing addresses', testData: { data: {} } },
-    ]
-
-    requests.forEach((req) => {
-      it(`${req.name}`, async () => {
-        try {
-          await execute(req.testData as AdapterRequest<TInputParameters>, {})
-        } catch (error) {
-          const errorResp = Requester.errored(jobID, error as AdapterError)
-          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
-        }
-      })
-    })
-  })
-})
